Fixup and more renames
