+++ packages/esm-patient-chart-app/src/loader/loader.component.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './loader.component.scss';
+import styles from './loader.scss';
 import { InlineLoading } from 'carbon-components-react';
 import { useTranslation } from 'react-i18next';
