+++ packages/esm-patient-chart-app/src/loader/loader.scss
