+++ packages/esm-patient-chart-app/src/patient-chart/action-menu/action-menu.scss
@@ -1,6 +1,6 @@
 @import "~@openmrs/esm-styleguide/src/vars";
 @import "~carbon-components/src/globals/scss/vars";
-@import "../root.scss";
+@import "../../root.scss";
 
 $icon-button-size: 48px;
 $actionPanelOffset: 48px;
