+++ packages/esm-patient-chart-app/src/patient-chart/patient-chart.scss
