+++ packages/esm-patient-chart-app/src/side-nav/side-menu.scss
