+++ packages/esm-patient-chart-app/src/visit/visit-form/visit-type-overview.scss
