+++ packages/esm-patient-chart-app/src/visit/visit-prompt/end-visit.scss
