+++ packages/esm-patient-chart-app/src/visit/visit-prompt/start-visit.scss
