+++ packages/esm-patient-chart-app/src/workspace/workspace-notification.component.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { cancelPrompt, useWorkspaces } from '@openmrs/esm-patient-common-lib';
 import { Button, ComposedModal, ModalBody, ModalFooter, ModalHeader } from 'carbon-components-react';
 import { useTranslation } from 'react-i18next';
-import styles from './workspace-notification.component.scss';
+import styles from './workspace-notification.scss';
 
 // t('activeFormWarning', 'There is an active form open in the workspace')
 // t('workspaceModalText', 'Launching a new form in the workspace could cause you to lose unsaved work on the {formName} form.')
