+++ packages/esm-patient-chart-app/src/workspace/workspace-notification.scss
